requirements.txt
websockets==10.2
pycryptodome==3.14.1
uvicorn==0.17.5
fastapi==0.75.0